Vars can provide DOM context
The notion of using vars to acquire DOM content is interesting. At first
I considered using vanilla JS but realized that I will always opt for
JQuery when I’m interacting with the DOM. So I’m not going to be so
dogmatic here. I struggled a little with scope and how to create a new node in the DOM
I needed to use `attr()` and then was able to chain together my DOM
commands.